# internal/docs/swagger.yaml
basePath: /v0
definitions:
  github_com_modelcontextprotocol_registry_internal_model.Argument:
    properties:
      choices:
        items:
          type: string
        type: array
      default:
        type: string
      description:
        type: string
      format:
        $ref: '#/definitions/github_com_modelcontextprotocol_registry_internal_model.Format'
      is_repeated:
        type: boolean
      is_required:
        type: boolean
      is_secret:
        type: boolean
      name:
        type: string
      properties:
        additionalProperties:
          $ref: '#/definitions/github_com_modelcontextprotocol_registry_internal_model.Input'
        type: object
      template:
        type: string
      type:
        $ref: '#/definitions/github_com_modelcontextprotocol_registry_internal_model.ArgumentType'
      value:
        type: string
      value_hint:
        type: string
      variables:
        additionalProperties:
          $ref: '#/definitions/github_com_modelcontextprotocol_registry_internal_model.Input'
        type: object
    type: object
  github_com_modelcontextprotocol_registry_internal_model.ArgumentType:
    enum:
    - positional
    - named
    type: string
    x-enum-varnames:
    - ArgumentTypePositional
    - ArgumentTypeNamed
  github_com_modelcontextprotocol_registry_internal_model.Format:
    enum:
    - string
    - number
    - boolean
    - file_path
    type: string
    x-enum-varnames:
    - FormatString
    - FormatNumber
    - FormatBoolean
    - FormatFilePath
  github_com_modelcontextprotocol_registry_internal_model.Input:
    properties:
      choices:
        items:
          type: string
        type: array
      default:
        type: string
      description:
        type: string
      format:
        $ref: '#/definitions/github_com_modelcontextprotocol_registry_internal_model.Format'
      is_required:
        type: boolean
      is_secret:
        type: boolean
      properties:
        additionalProperties:
          $ref: '#/definitions/github_com_modelcontextprotocol_registry_internal_model.Input'
        type: object
      template:
        type: string
      value:
        type: string
    type: object
  github_com_modelcontextprotocol_registry_internal_model.KeyValueInput:
    properties:
      choices:
        items:
          type: string
        type: array
      default:
        type: string
      description:
        type: string
      format:
        $ref: '#/definitions/github_com_modelcontextprotocol_registry_internal_model.Format'
      is_required:
        type: boolean
      is_secret:
        type: boolean
      name:
        type: string
      properties:
        additionalProperties:
          $ref: '#/definitions/github_com_modelcontextprotocol_registry_internal_model.Input'
        type: object
      template:
        type: string
      value:
        type: string
      variables:
        additionalProperties:
          $ref: '#/definitions/github_com_modelcontextprotocol_registry_internal_model.Input'
        type: object
    type: object
  github_com_modelcontextprotocol_registry_internal_model.Package:
    properties:
      environment_variables:
        items:
          $ref: '#/definitions/github_com_modelcontextprotocol_registry_internal_model.KeyValueInput'
        type: array
      name:
        type: string
      package_arguments:
        items:
          $ref: '#/definitions/github_com_modelcontextprotocol_registry_internal_model.Argument'
        type: array
      registry_name:
        type: string
      runtime_arguments:
        items:
          $ref: '#/definitions/github_com_modelcontextprotocol_registry_internal_model.Argument'
        type: array
      runtime_hint:
        type: string
      version:
        type: string
    type: object
  github_com_modelcontextprotocol_registry_internal_model.PublishRequest:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
      packages:
        items:
          $ref: '#/definitions/github_com_modelcontextprotocol_registry_internal_model.Package'
        type: array
      remotes:
        items:
          $ref: '#/definitions/github_com_modelcontextprotocol_registry_internal_model.Remote'
        type: array
      repository:
        $ref: '#/definitions/github_com_modelcontextprotocol_registry_internal_model.Repository'
      status:
        $ref: '#/definitions/github_com_modelcontextprotocol_registry_internal_model.ServerStatus'
      version_detail:
        $ref: '#/definitions/github_com_modelcontextprotocol_registry_internal_model.VersionDetail'
    type: object
  github_com_modelcontextprotocol_registry_internal_model.Remote:
    properties:
      headers:
        items:
          $ref: '#/definitions/github_com_modelcontextprotocol_registry_internal_model.Input'
        type: array
      transport_type:
        type: string
      url:
        type: string
    type: object
  github_com_modelcontextprotocol_registry_internal_model.Repository:
    properties:
      id:
        type: string
      source:
        type: string
      url:
        type: string
    type: object
  github_com_modelcontextprotocol_registry_internal_model.Server:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
      repository:
        $ref: '#/definitions/github_com_modelcontextprotocol_registry_internal_model.Repository'
      status:
        $ref: '#/definitions/github_com_modelcontextprotocol_registry_internal_model.ServerStatus'
      version_detail:
        $ref: '#/definitions/github_com_modelcontextprotocol_registry_internal_model.VersionDetail'
    type: object
  github_com_modelcontextprotocol_registry_internal_model.ServerStatus:
    enum:
    - active
    - deprecated
    type: string
    x-enum-varnames:
    - ServerStatusActive
    - ServerStatusDeprecated
  github_com_modelcontextprotocol_registry_internal_model.VersionDetail:
    properties:
      is_latest:
        type: boolean
      release_date:
        type: string
      version:
        type: string
    type: object
  internal_api_handlers_v0.HealthResponse:
    properties:
      github_client_id:
        type: string
      status:
        type: string
    type: object
  internal_api_handlers_v0.Metadata:
    properties:
      count:
        type: integer
      next_cursor:
        type: string
      total:
        type: integer
    type: object
  internal_api_handlers_v0.PaginatedResponse:
    properties:
      metadata:
        $ref: '#/definitions/internal_api_handlers_v0.Metadata'
      servers:
        items:
          $ref: '#/definitions/github_com_modelcontextprotocol_registry_internal_model.Server'
        type: array
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A community driven registry service for Model Context Protocol (MCP)
    servers.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: MCP Registry API
  version: "1.0"
paths:
  /health:
    get:
      description: Check the health status of the API
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handlers_v0.HealthResponse'
        "500":
          description: Failed to encode response
          schema:
            type: string
      summary: Health check
      tags:
      - health
  /publish:
    post:
      consumes:
      - application/json
      description: Publish a new MCP server to the registry or update an existing
        one
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Server details
        in: body
        name: server
        required: true
        schema:
          $ref: '#/definitions/github_com_modelcontextprotocol_registry_internal_model.PublishRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_modelcontextprotocol_registry_internal_model.Server'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "405":
          description: Method not allowed
          schema:
            type: string
        "500":
          description: Failed to publish server
          schema:
            type: string
      security:
      - Bearer: []
      summary: Publish MCP server
      tags:
      - publish
  /servers:
    get:
      consumes:
      - application/json
      description: Get a paginated list of MCP servers from the registry
      parameters:
      - description: Pagination cursor (UUID)
        in: query
        name: cursor
        type: string
      - description: 'Number of items per page (default: 30)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handlers_v0.PaginatedResponse'
        "400":
          description: Invalid cursor parameter
          schema:
            type: string
        "405":
          description: Method not allowed
          schema:
            type: string
        "500":
          description: Failed to get registry list
          schema:
            type: string
      summary: List MCP servers
      tags:
      - servers
  /servers/{id}:
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific MCP server
      parameters:
      - description: Server ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_modelcontextprotocol_registry_internal_model.Server'
        "400":
          description: Invalid server ID format
          schema:
            type: string
        "404":
          description: Server not found
          schema:
            type: string
        "405":
          description: Method not allowed
          schema:
            type: string
        "500":
          description: Failed to get server details
          schema:
            type: string
      summary: Get MCP server details
      tags:
      - servers
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and GitHub token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"